remove previous job system
