+++ convex/_generated/api.d.ts
@@ -29,27 +29,20 @@ import type * as endpoints_openai from "../endpoints/openai.js";
 import type * as endpoints_openrouter from "../endpoints/openrouter.js";
 import type * as endpoints_sinkin from "../endpoints/sinkin.js";
 import type * as endpoints_together from "../endpoints/together.js";
-import type * as files_createImageFromUrl from "../files/createImageFromUrl.js";
 import type * as files_ingestImageUrl from "../files/ingestImageUrl.js";
 import type * as files_links from "../files/links.js";
-import type * as files_optimizeImageFile from "../files/optimizeImageFile.js";
 import type * as functions from "../functions.js";
 import type * as http from "../http.js";
 import type * as images from "../images.js";
 import type * as inference_assessNsfw from "../inference/assessNsfw.js";
 import type * as inference_captionImage from "../inference/captionImage.js";
 import type * as inference_chat from "../inference/chat.js";
-import type * as inference_chatCompletion from "../inference/chatCompletion.js";
 import type * as inference_chatCompletionAi from "../inference/chatCompletionAi.js";
-import type * as inference_soundGeneration from "../inference/soundGeneration.js";
 import type * as inference_textToAudio from "../inference/textToAudio.js";
 import type * as inference_textToImage from "../inference/textToImage.js";
-import type * as inference_textToImageNext from "../inference/textToImageNext.js";
 import type * as inference_textToSpeech from "../inference/textToSpeech.js";
 import type * as inference_threadTitleCompletion from "../inference/threadTitleCompletion.js";
-import type * as inference_threadTitleCompletionNext from "../inference/threadTitleCompletionNext.js";
 import type * as jobs from "../jobs.js";
-import type * as jobsNext from "../jobsNext.js";
 import type * as lib_clerk from "../lib/clerk.js";
 import type * as lib_openai from "../lib/openai.js";
 import type * as lib_paramInputs from "../lib/paramInputs.js";
@@ -97,27 +90,20 @@ declare const fullApi: ApiFromModules<{
   "endpoints/openrouter": typeof endpoints_openrouter;
   "endpoints/sinkin": typeof endpoints_sinkin;
   "endpoints/together": typeof endpoints_together;
-  "files/createImageFromUrl": typeof files_createImageFromUrl;
   "files/ingestImageUrl": typeof files_ingestImageUrl;
   "files/links": typeof files_links;
-  "files/optimizeImageFile": typeof files_optimizeImageFile;
   functions: typeof functions;
   http: typeof http;
   images: typeof images;
   "inference/assessNsfw": typeof inference_assessNsfw;
   "inference/captionImage": typeof inference_captionImage;
   "inference/chat": typeof inference_chat;
-  "inference/chatCompletion": typeof inference_chatCompletion;
   "inference/chatCompletionAi": typeof inference_chatCompletionAi;
-  "inference/soundGeneration": typeof inference_soundGeneration;
   "inference/textToAudio": typeof inference_textToAudio;
   "inference/textToImage": typeof inference_textToImage;
-  "inference/textToImageNext": typeof inference_textToImageNext;
   "inference/textToSpeech": typeof inference_textToSpeech;
   "inference/threadTitleCompletion": typeof inference_threadTitleCompletion;
-  "inference/threadTitleCompletionNext": typeof inference_threadTitleCompletionNext;
   jobs: typeof jobs;
-  jobsNext: typeof jobsNext;
   "lib/clerk": typeof lib_clerk;
   "lib/openai": typeof lib_openai;
   "lib/paramInputs": typeof lib_paramInputs;
