+++ convex/db/threads.ts
@@ -4,8 +4,7 @@ import { v } from 'convex/values'
 
 import { internal } from '../_generated/api'
 import { mutation, query } from '../functions'
-import { createJob } from '../jobs'
-import { createJob as createJobNext } from '../jobsNext'
+import { createJob as createJobNext } from '../jobs'
 import { inferenceConfigV, threadFields } from '../schema'
 import { defaultChatInferenceConfig } from '../shared/defaults'
 import { createError, insist } from '../shared/utils'
