+++ convex/files/ingestImageUrl.ts
@@ -2,7 +2,7 @@ import { v } from 'convex/values'
 
 import { internal } from '../_generated/api'
 import { internalAction, internalMutation } from '../functions'
-import { claimJob, completeJob } from '../jobsNext'
+import { claimJob, completeJob } from '../jobs'
 import { insist } from '../shared/utils'
 
 export const claim = internalMutation({
