+++ convex/inference/threadTitleCompletion.ts
@@ -1,42 +1,39 @@
+import { anthropic } from '@ai-sdk/anthropic'
+import { generateText } from 'ai'
 import { v } from 'convex/values'
-import { z } from 'zod'
 
 import { internal } from '../_generated/api'
 import { internalAction, internalMutation } from '../functions'
-import { acquireJob, handleJobError, jobResultSuccess } from '../jobs'
-import { createOpenAiClient } from '../lib/openai'
+import { claimJob, completeJob, handleJobError } from '../jobs'
 import { insist } from '../shared/utils'
 
-const msgSchema = z.object({
-  role: z.enum(['system', 'assistant', 'user']),
-  name: z.string().optional(),
-  content: z.string(),
-})
-
-export const init = internalMutation({
+export const start = internalMutation({
   args: {
     jobId: v.id('jobs'),
   },
   handler: async (ctx, args) => {
-    const job = await acquireJob(ctx, args.jobId)
-    const threadId = job.threadId
-    insist(threadId, 'no threadId', { code: 'invalid_job_input' })
+    const job = await claimJob(ctx, args)
+    insist(job.threadId, 'required: threadId', { code: 'invalid_job' })
 
-    const messages: z.infer<typeof msgSchema>[] = await ctx
-      .table('messages', 'threadId', (q) => q.eq('threadId', threadId))
+    const thread = await ctx.table('threads').getX(job.threadId)
+    const messages = await ctx
+      .table('messages', 'threadId', (q) => q.eq('threadId', thread._id))
       .order('desc')
       .filter((q) =>
         q.and(
           q.eq(q.field('deletionTime'), undefined),
           q.neq(q.field('content'), undefined),
           q.neq(q.field('role'), 'system'),
-          q.neq(q.field('content'), undefined),
         ),
       )
       .take(4)
-      .map((message) => msgSchema.parse(message))
+    insist(messages.length > 0, 'required: messages', { code: 'invalid_job' })
 
-    return { threadId, messages: messages.toReversed() }
+    return {
+      job,
+      threadId: thread._id,
+      messages,
+    }
   },
 })
 
@@ -44,46 +41,49 @@ export const run = internalAction({
   args: {
     jobId: v.id('jobs'),
   },
-  handler: async (ctx, { jobId }) => {
+  handler: async (ctx, args) => {
     try {
-      const input = await ctx.runMutation(internal.inference.threadTitleCompletion.init, { jobId })
+      const { job, threadId, messages } = await ctx.runMutation(
+        internal.inference.threadTitleCompletion.start,
+        args,
+      )
 
-      const api = createOpenAiClient(endpoint)
-
-      const message = titleMessage.replace(
+      const message = prompt.replace(
         '%%%',
-        input.messages
+        messages
           .map((message) => message.content)
           .join('\n')
           .slice(0, 500),
       )
 
-      const body = {
-        messages: [
-          {
-            role: 'system' as const,
-            content: message,
+      console.log('[threadTitleCompletion] [input] [anthropic]', message)
+      const response = await generateText({
+        model: anthropic('claude-3-haiku-20240307'),
+        prompt: message,
+        maxTokens: 1024,
+      })
+      console.log('[threadTitleCompletion] [output] [anthropic]', response)
+
+      const title = response.text.split('\n').at(-1)
+      if (!title) {
+        await ctx.runMutation(internal.jobs.fail, {
+          jobId: job._id,
+          jobError: {
+            message: 'no title',
+            code: 'invalid_response',
+            fatal: false,
           },
-        ],
-        max_tokens: 1024,
-        model: 'meta-llama/Llama-3-8b-chat-hf',
-        stream: false as const,
+        })
+        return
       }
-      console.log('title-completion', 'together', body)
-
-      const chatCompletion = await api.chat.completions.create(body)
-      console.log(chatCompletion)
-
-      const content = chatCompletion.choices[0]?.message.content ?? '' //TODO check
-      const title = content.split('\n').at(-1) ?? '<title?>'
 
       await ctx.runMutation(internal.inference.threadTitleCompletion.complete, {
-        jobId,
-        threadId: input.threadId,
+        jobId: job._id,
+        threadId,
         title,
       })
-    } catch (err) {
-      return await handleJobError(ctx, { err, jobId })
+    } catch (error) {
+      await handleJobError(ctx, { jobId: args.jobId, error })
     }
   },
 })
@@ -94,14 +94,13 @@ export const complete = internalMutation({
     threadId: v.id('threads'),
     title: v.string(),
   },
-  handler: async (ctx, { jobId, threadId, title }) => {
-    await ctx.skipRules.table('threads').getX(threadId).patch({ title })
-    await jobResultSuccess(ctx, { jobId })
+  handler: async (ctx, args) => {
+    await ctx.skipRules.table('threads').getX(args.threadId).patch({ title: args.title })
+    await completeJob(ctx, args)
   },
 })
 
-const endpoint = 'together'
-const titleMessage = `
+const prompt = `
 The following messages are from the start of a conversation that a user has just initiated with
 another large language model. Your task is to create a succinct and concise summary of the
 conversation so far, and create a title for the conversation, which is no more than 6 words.
